11-audio-analysis/requirements.txt
assemblyai==0.41.5
mcp==1.9.4
python-dotenv==1.1.1
streamlit==1.37.1
